move startup function to setup plugin
